refactor: extract shared user form
